% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{concept_nc}
\alias{concept_nc}
\alias{concept_nc.tricontext.data.table}
\title{Unchecked concept generation mathods}
\usage{
concept_nc(x, ...)

\method{concept_nc}{tricontext.data.table}(cxt, one, two, names)
}
\arguments{
\item{cxt}{triadic formal context}

\item{one}{a data.table of elements of one of the three defining sets}

\item{two}{a data.table of elements of another of the three defining sets}

\item{names}{a vector containing the three names of the three defining sets
The first one corresponds to the parameter \code{one}, the secound one to \code{two}, and the third one to the missing dimension.}
}
\value{
triadic concept
}
\description{
Generic method for creating formal concepts (unchecked version)
}
\section{Methods (by class)}{
\itemize{
\item \code{tricontext.data.table}: Create a formal triadic concept from a triadic
extent and a triadic intent
}}
\section{Warning}{
 This functions should be called after checking
all parameters, only.  Otherwise, expect cryptic error
messages. For debugging purpose try to use \code{\link{concept}}
instead.
}
\seealso{
concept

\code{\link{concept}}
}

