% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{extent}
\alias{extent}
\alias{extent.tricontext.data.table}
\alias{extent.triextent.data.table}
\alias{extent.triintent.data.table}
\alias{extent.trimodintent.data.table}
\alias{extent.trimodus.data.table}
\title{Create an extent from some object.}
\usage{
extent(x, ...)

\method{extent}{triextent.data.table}(extent)

\method{extent}{tricontext.data.table}(context, objects)

\method{extent}{trimodus.data.table}(modus, intent)

\method{extent}{triintent.data.table}(intent, modus)

\method{extent}{trimodintent.data.table}(x)
}
\arguments{
\item{x}{description of the object
A see extent.tricontext.data.table for information about creating an
extent from a list of objects.}

\item{context}{formal tricontext, the possible extent should be
associated with}

\item{objects}{a subset of objects of the formal tricontext context}

\item{modus}{a triadic modus}

\item{intent}{a triadic intent}

\item{intent}{a triadic intent}

\item{modus}{a triadic modus}

\item{x}{a trimodintent as generated by modintent.}
}
\value{
a formal triadic extent

a formal triextent

a triadic intent.

a triadic intent.

a triadic extent representing the objects that
are shared by all condition/attribute pairs of the triadic modintent x
}
\description{
Create an extent from some object.
}
\section{Methods (by class)}{
\itemize{
\item \code{triextent.data.table}: return the extent as no conversion is needed

\item \code{tricontext.data.table}: Convert a list of object names into a possible formal extent of a
triadic context

In fact this is a lazy definiton of an extent. The object set is
not necessarily the extent of a formal triconcept. This function
creates the link between a set of object names and a formal
tricontext.

\item \code{trimodus.data.table}: Get the attributes from a formal triadic context that are shared by
all elemets of the set theoretical product of a modus with an
intent.

The result is the set of all attributes that are connected to all
given objects under all given conditions via the incidence relation.

\item \code{triintent.data.table}: Get the attributes from a formal triadic context that are shared by
all elemets of the set theoretical product of a modus with an
intent.

The result is the set of all attributes that are connected to all
given objects under all given conditions via the incidence relation.

\item \code{trimodintent.data.table}: Get the objects from a formal triadic context that are
shared by a set of given condition/attribute pairs.

The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of objects and the
product of the condition and attribute sets that is derived from the
triadic context.
}}
\seealso{
extent.tricontext.data.table

Other FCA.derivation: \code{\link{intent}}
}

