% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{extent.tricontext.data.table}
\alias{extent.tricontext.data.table}
\title{Convert a list of object names into a possible formal extent of a
triadic context}
\usage{
\method{extent}{tricontext.data.table}(context, objects)
}
\arguments{
\item{context}{formal tricontext, the possible extent should be
associated with}

\item{objects}{a subset of objects of the formal tricontext context}
}
\value{
a formal triextent
}
\description{
In fact this is a lazy definiton of an extent. The object set is
not necessarily the extent of a formal triconcept. This function
creates the link between a set of object names and a formal
tricontext.
}

