% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{extent.trimodintent.data.table}
\alias{extent.trimodintent.data.table}
\title{Get the objects from a formal triadic context that are
shared by a set of given condition/attribute pairs.}
\usage{
\method{extent}{trimodintent.data.table}(x)
}
\arguments{
\item{x}{a trimodintent as generated by modintent.}
}
\value{
a triadic extent representing the objects that
are shared by all condition/attribute pairs of the triadic modintent x
}
\description{
The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of objects and the
product of the condition and attribute sets that is derived from the
triadic context.
}

