% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{extintent}
\alias{extintent}
\title{Get a set of object/attribute pairs depending on a givon R object}
\usage{
extintent(x, ...)
}
\arguments{
\item{x}{description of the object
A see extintent.trimodus.data.table for information about creating an
a partial triadic derivation of an modus.}
}
\value{
a an object representing object/attribute pairs
}
\description{
Get a set of object/attribute pairs depending on a givon R object
}

