% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{extintent.trimodus.data.table}
\alias{extintent.trimodus.data.table}
\title{Get the object/attribute pairs from a formal triadic context that are
shared by a set of given attributes.}
\usage{
\method{extintent}{trimodus.data.table}(x)
}
\arguments{
\item{x}{a triadic modus.}
}
\value{
a an object representing the object/attribute pairs that
are shared by all attributes of the triadic modus x
}
\description{
The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the object and attribute sets that is derived from the
triadic context.
}

