% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{intent}
\alias{intent}
\alias{intent.tricontext.data.table}
\alias{intent.triextent.data.table}
\alias{intent.triintent.data.table}
\alias{intent.trimodextent.data.table}
\alias{intent.trimodus.data.table}
\title{Create an intent from some object.}
\usage{
intent(x, ...)

\method{intent}{triintent.data.table}(intent)

\method{intent}{tricontext.data.table}(context, attributes)

\method{intent}{trimodextent.data.table}(x)

\method{intent}{trimodus.data.table}(modus, extent)

\method{intent}{triextent.data.table}(extent, modus)
}
\arguments{
\item{x}{description of the object
A see intent.tricontext.data.table for information about creating an
intent from a list of objects.}

\item{context}{formal tricontext, the possible intent should be
associated with}

\item{attributes}{a subset of attributes of the formal tricontext context}

\item{modus}{a triadic modus}

\item{extent}{a triadic extent}

\item{x}{a trimodextent as generated by modextent.}

\item{extent}{a triadic extent}

\item{modus}{a triadic modus}
}
\value{
a formal triadic intent

a formal triintent

a triadic intent representing the attributes that
are shared by all condition/object pairs of the triadic modextent x

a triadic intent.

a triadic intent.
}
\description{
Create an intent from some object.
}
\section{Methods (by class)}{
\itemize{
\item \code{triintent.data.table}: return the intent as no conversion is needed

\item \code{tricontext.data.table}: Convert a list of attribute names into a possible formal intent of a
triadic context

In fact this is a lazy definiton of an intent. The attribute set is
not necessarily the intent of a formal triconcept. This function
creates the link between a set of attribute names and a formal
tricontext.

\item \code{trimodextent.data.table}: Get the attributes from a formal triadic context that are
shared by a set of given condition/object pairs.

The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the condition and object sets that is derived from the
triadic context.

\item \code{trimodus.data.table}: Get the attributes from a formal triadic context that are shared by
all elemets of the set theoretical product of a modus with an
extent.

The result is the set of all attributes that are connected to all
given objects under all given conditions via the incidence relation.

\item \code{triextent.data.table}: Get the attributes from a formal triadic context that are shared by
all elemets of the set theoretical product of a modus with an
extent.

The result is the set of all attributes that are connected to all
given objects under all given conditions via the incidence relation.
}}
\seealso{
extent.tricontext.data.table

Other FCA.derivation: \code{\link{extent}}
}

