% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{intent.tricontext.data.table}
\alias{intent.tricontext.data.table}
\title{Convert a list of attribute names into a possible formal intent of a
triadic context}
\usage{
\method{intent}{tricontext.data.table}(context, attributes)
}
\arguments{
\item{context}{formal tricontext, the possible intent should be
associated with}

\item{attributes}{a subset of attributes of the formal tricontext context}
}
\value{
a formal triintent
}
\description{
In fact this is a lazy definiton of an intent. The attribute set is
not necessarily the intent of a formal triconcept. This function
creates the link between a set of attribute names and a formal
tricontext.
}

