% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{intent.trimodextent.data.table}
\alias{intent.trimodextent.data.table}
\title{Get the attributes from a formal triadic context that are
shared by a set of given condition/object pairs.}
\usage{
\method{intent}{trimodextent.data.table}(x)
}
\arguments{
\item{x}{a trimodextent as generated by modextent.}
}
\value{
a triadic intent representing the attributes that
are shared by all condition/object pairs of the triadic modextent x
}
\description{
The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the condition and object sets that is derived from the
triadic context.
}

