% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{is.tricontext.data.table}
\alias{is.tricontext.data.table}
\title{tests whether an object is a formal tricontext}
\usage{
is.tricontext.data.table(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
TRUE, if the object is a formal tricontext
FALSE, otherwise
}
\description{
tests whether an object is a formal tricontext
}

