% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modextent.triintent.data.table}
\alias{modextent.triintent.data.table}
\title{Get the condition/object pairs from a formal triadic context that are
shared by a set of given attributes.}
\usage{
\method{modextent}{triintent.data.table}(x)
}
\arguments{
\item{x}{a triadic intent.}
}
\value{
a an object representing the condition/object pairs that
are shared by all attributes of the triadic intent x
}
\description{
The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the condition and object sets that is derived from the
triadic context.
}

