% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modintent}
\alias{modintent}
\title{Get a set of condition/attribute pairs depending on a givon R object}
\usage{
modintent(x, ...)
}
\arguments{
\item{x}{description of the object
A see modintent.triintent.data.table for information about creating an
a partial triadic derivation of an extent.}
}
\value{
a an object representing condition/attribute pairs
}
\description{
Get a set of condition/attribute pairs depending on a givon R object
}

