% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modintent.triextent.data.table}
\alias{modintent.triextent.data.table}
\title{Get the condition/attribute pairs from a formal triadic context that are
shared by a set of given attributes.}
\usage{
\method{modintent}{triextent.data.table}(x)
}
\arguments{
\item{x}{a triadic extent.}
}
\value{
a an object representing the condition/attribute pairs that
are shared by all attributes of the triadic extent x
}
\description{
The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the condition and attribute sets that is derived from the
triadic context.
}

