% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modus}
\alias{modus}
\alias{modus.tricontext.data.table}
\alias{modus.triextent.data.table}
\alias{modus.triextintent.data.table}
\alias{modus.triintent.data.table}
\alias{modus.trimodus.data.table}
\title{Create an modus from some object.}
\usage{
modus(x, ...)

\method{modus}{trimodus.data.table}(modus)

\method{modus}{tricontext.data.table}(context, conditions)

\method{modus}{triextent.data.table}(extent, intent)

\method{modus}{triintent.data.table}(intent, extent)

\method{modus}{triextintent.data.table}(x)
}
\arguments{
\item{x}{description of the object
A see modus.tricontext.data.table for information about creating an
modus from a list of objects.}

\item{context}{formal tricontext, the possible modus should be
associated with}

\item{conditions}{a subset of conditions of the formal tricontext context}

\item{extent}{a triadic extent}

\item{intent}{a triadic intent}

\item{intent}{a triadic intent}

\item{extent}{a triadic extent}

\item{x}{a triextintent as generated by extintent.}
}
\value{
a formal triadic modus

a formal trimodus

a triadic intent.

a triadic intent.

a modus representing the conditions that
are shared by all object/attribute pairs of the triadic extintent x
}
\description{
Create an modus from some object.
}
\section{Methods (by class)}{
\itemize{
\item \code{trimodus.data.table}: return the modus as no conversion is needed

\item \code{tricontext.data.table}: Convert a list of condition names into a possible formal modus of a
triadic context

In fact this is a lazy definiton of an modus. The condition set is
not necessarily the modus of a formal triconcept. This function
creates the link between a set of condition names and a formal
tricontext.

\item \code{triextent.data.table}: Get the attributes from a formal triadic context that are shared by
all elemets of the set theoretical product of a extent with an
intent.

The result is the set of all attributes that are connected to all
given objects under all given conditions via the incidence relation.

\item \code{triintent.data.table}: Get the attributes from a formal triadic context that are shared by
all elemets of the set theoretical product of a extent with an
intent.

The result is the set of all attributes that are connected to all
given objects under all given conditions via the incidence relation.

\item \code{triextintent.data.table}: Get the conditions from a formal triadic context that are
shared by a set of given object/attribute pairs.

The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the object and attribute sets that is derived from the
triadic context.
}}

