% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modus.tricontext.data.table}
\alias{modus.tricontext.data.table}
\title{Convert a list of condition names into a possible formal modus of a
triadic context}
\usage{
\method{modus}{tricontext.data.table}(context, conditions)
}
\arguments{
\item{context}{formal tricontext, the possible modus should be
associated with}

\item{conditions}{a subset of conditions of the formal tricontext context}
}
\value{
a formal trimodus
}
\description{
In fact this is a lazy definiton of an modus. The condition set is
not necessarily the modus of a formal triconcept. This function
creates the link between a set of condition names and a formal
tricontext.
}

