% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modus.triextintent.data.table}
\alias{modus.triextintent.data.table}
\title{Get the conditions from a formal triadic context that are
shared by a set of given object/attribute pairs.}
\usage{
\method{modus}{triextintent.data.table}(x)
}
\arguments{
\item{x}{a triextintent as generated by extintent.}
}
\value{
a modus representing the conditions that
are shared by all object/attribute pairs of the triadic extintent x
}
\description{
The result is a triadic equivalent to the binary derivation
operation in the (binary) formal context of attributes and the
product of the object and attribute sets that is derived from the
triadic context.
}

