% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{concept}
\alias{concept}
\alias{concept.triconcept.data.table}
\title{Create an extent from some object.}
\usage{
concept(x, ...)

\method{concept}{triconcept.data.table}(concept)

concept(x, ...)
}
\arguments{
\item{x}{description of the object
A see extent.tricontext.data.table for information about creating an
extent from a list of objects.}
}
\value{
a formal triadic extent
}
\description{
Create an extent from some object.

Create a formal concept
}
\section{Methods (by class)}{
\itemize{
\item \code{triconcept.data.table}: return the concept as no conversion is needed
}}

