% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{intent}
\alias{intent}
\title{Create an intent from some object.}
\usage{
intent(x, ...)
}
\arguments{
\item{x}{description of the object
A see intent.tricontext.data.table for information about creating an
intent from a list of objects.}
}
\value{
a formal triadic intent
}
\description{
Create an intent from some object.
}

