% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modextent}
\alias{modextent}
\title{Get a set of condition/object pairs depending on a givon R object}
\usage{
modextent(x, ...)
}
\arguments{
\item{x}{description of the object
A see modextent.triintent.data.table for information about creating an
a partial triadic derivation of an intent.}
}
\value{
a an object representing condition/object pairs
}
\description{
Get a set of condition/object pairs depending on a givon R object
}

