% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tricontext.R
\name{modus}
\alias{modus}
\title{Create an modus from some object.}
\usage{
modus(x, ...)
}
\arguments{
\item{x}{description of the object
A see modus.tricontext.data.table for information about creating an
modus from a list of objects.}
}
\value{
a formal triadic modus
}
\description{
Create an modus from some object.
}

